requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
google-api-core==2.2.1
google-auth==2.3.2
google-cloud-speech==2.11.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.4
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
packaging==21.2
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyparsing==2.4.7
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
